+++ client/src/editors/httpprovider.ts
@@ -1,31 +1,6 @@
 import { CustomTextEditorProvider, WebviewPanel, TextDocument, Webview, Uri, ExtensionContext, window } from "vscode";
-import { parse } from "fast-xml-parser"
 import path from "path";
-
-interface HttpService {
-    handlerClass: string,
-    author: string,
-    name: string,
-    text: string,
-    url: string,
-}
-
-export const parseHTTP = (s: string): HttpService => {
-    const raw = parse(s, {
-        ignoreAttributes: false,
-        trimValues: false,
-        parseAttributeValue: true
-    })
-    const {
-        "@_http:handler_servicehandler": handlerClass = "",
-        "@_adtcore:responsible": author = "",
-        "@_adtcore:name": name = "",
-        "@_adtcore:description": text = "",
-        "@_http:header_canonicalurl": url = "", } = raw["http:abap"] || {}
-
-    return { handlerClass, author, name, text, url }
-}
-
+import { parseHTTP } from "./httpparser";
 
 export class HttpProvider implements CustomTextEditorProvider {
     constructor(private context: ExtensionContext) { }