+++ modules/abapObject/jest.config.js
@@ -1,5 +1,7 @@
 // For a detailed explanation regarding each configuration property, visit:
 // https://jestjs.io/docs/en/configuration.html
+const { existsSync } = require("fs")
+const setupFiles = ["./setenv.js"].filter(existsSync)
 
 module.exports = {
   // A set of global variables that need to be available in all test environments
@@ -16,7 +18,7 @@ module.exports = {
   moduleFileExtensions: ["ts", "tsx", "js"],
 
   // The paths to modules that run some code to configure or set up the testing environment before each test
-  setupFiles: ["<rootDir>/setenv.js"],
+  setupFiles,
 
   // The test environment that will be used for testing
   testEnvironment: "node",